[YPUB-6537] Fix (Long->Integer)

+++ src/eu/ydp/empiria/player/client/util/events/dom/emulate/events/pointer/PointerNativeMethods.java
@@ -15,7 +15,7 @@ public class PointerNativeMethods {
 		return elem.isPrimary;
 	}-*/;
 
-	public native Long getPointerId(NativeEvent elem) /*-{
+	public native Integer getPointerId(NativeEvent elem) /*-{
 		return elem.pointerId;
 	}-*/;
 }
